requirements.txt
discord.py
fastapi
langchain
langchain-openai
openai
python-dotenv
redis-py-cluster
ruff
uvicorn
